services/spotify_service/requirements.txt
python-dotenv
pyyaml
fastapi
uvicorn
httpx
pydantic_extra_types
pytest
pytest-asyncio
pytest-mock
